Move ACAlgorithm implementation files
Separating ACAlgorithm implementation files by moving them into
/algebraic_constraints directory
