+++ src/algorithms/legacy_algorithms.h
@@ -1,4 +1,4 @@
 #pragma once
 
 /* Algebraic constraints mining algorithm */
-#include "algorithms/ac_algorithm.h"
+#include "algorithms/algebraic_constraints/ac_algorithm.h"
